Implement token bucket algorithm
A token bucket is a way to keep track of a rate of events over time, but
allowing for bursts of the events. This is done by having a pool of
tokens that are consumed by each event, and then regenerated over time.

The token bucket class holds the tokens available, the rate by which tokens
should be regenerated, the maximum number of tokens the bucket should hold and
lastly the Backend where the bucket should persist itself.

When a token bucket is cached its total tokens are stored along with the last
consume timestamp. From these two pieces of information the current number
of tokens can be calculated.

The TokenBucket is only cached for the time it takes the bucket to reach
the max number of tokens. After that point, there is no reason to keep
its state, since it's implied that it will have the maximum number of
tokens.

Since the token bucket uses integers to store its tokens, it is limited to
PHP_INT_MAX number of tokens.