Make custom interface for backend
Instead of adding Matryoshka as a dependency for the sake of its Backend
interface, create a custom one for the use of storing token buckets.
This will remove the need to have extra dependencies.

The abstract interface requires a StoredBucket instead of a generic
value so that there can more strict enforcement of types.

Some small problems with the date format string don't allow for larger
than a 32 bit signed integer maximum. Fixed in readyTime where tests
were pushing it over the readyTime

Uses a namespace now! Also, all of the files need to be loaded.
Autoloading seemed overkill since there are not that many.