Return timestamp in public interface
It was pointed out that it is easier and preferred to work with a
timestamp for an interface rather than a DateTime object.

Internally, the storage stores the last consume time as a timestamp, so
working with a timestamp on the public interface shoulnd't add any
conocerns that didn't already exist.

Updates tests to receive the timestamp and convert it back into a
DateTime for its assertion.