Convert TokenRate to use seconds as integers
There is is a flaw with using DateTimes and DateIntervals to do
arithmetic. A DateInterval won't actually convert itself to entirely
seconds if you try to get the seconds value of a DateInterval that has
more than just seconds.

This is a big problem for the use of the class for TokenBucket because
we use the seconds to calculate everything.