Convert TokenBucket and StoredBucket rom DateTime
Moving away from the DateTime objects because there is no way to get
all seconds from DateIntervals.

Using doubles for the number of seconds so that microseconds can be
used.

Adding in more type assertions for parameters of functions even for
functions that aren't public. This will eventually be replaced by scalar
type hinting with php 7. Until then, this will fulfill the same goal.