Use Objects and remove helper
The helper isn't really needed and its name is a little confusing, since
it was the same as getTokens but with helper appended. Removing the
helper entirely and seperating the pieces into two seperate functions.

Using the StoredBucket to retrieve the readyTime. The ready time relies
on the current tokens, which is a property of the stored bucket.

Assert that all the returns are doubles. Since using just seconds can
allow us to work with microtime as well, then there isn't a reason to be
using an int. If someone using this wants an int, they can just cast it.