Calculate the correct ready time unable to consume
The amount of tokens to be used for the ready time calculation should
be the same as the stored bucket. If we aren't going to edit the stored
bucket since the consume failed, then the ready time should be using the
old amount of tokens, not the updated amount.

Updates comments to fix the change from DateTime to an numeric time.