Fix logic for consumption
Checks against the right amount that has the regenerated amount of
tokens. There wasn't a true way to test that a consume was actually
consuming from the regenerated tokens and not from the stored bucket.

Creates a new function getStoredBucket that gets the bucket for you, or
if it is not in cache, then it generates a fresh stored bucket.

Fixes offset to set the member variable correctly.

Correclty takes the maximum from the readyTime so it's not going below
0, not so it doesn't go over zero.

Adds a cool new test that tests that consume is working with the
regenerated amount.