Fix getReadyTime calculation
The get readyTime calculation was accidentally a calculation of tokens *
a rate instead of the number of seconds required to get the token
amount.

The ready time was also being returned as the current timestamp if the
consume was successful. This should now be 0 because it is ready now,
and the interval shouldn't be anything.