Token Rates: allow rates < 1 tok/sec
Allow rates like 5 tokens every 20 sec.

Instead of forcing everything to an int and
breaking all token rates<1, just do math.

If the consumer needs to consume them in discreet chunks, it can ask for
them as such ->consume(1). If the resource is continuous in nature,
the consumer can ask for ->consume(2.7556) and it will work as expected.

Fixes a bug with Rates < 1 and deals with Rates < 0 by simply short
cirtuiting the request and saying No.

Add tests for Rate < 1 and Rate == 0